+++ mea/src/once/once/mod.rs
@@ -20,6 +20,9 @@ use std::task::Poll;
 use crate::internal::CountdownState;
 use crate::semaphore::Semaphore;
 
+#[cfg(test)]
+mod tests;
+
 /// A synchronization primitive which can be used to run a one-time async initialization.
 ///
 /// Unlike [`std::sync::Once`], this type never blocks a thread. The provided closure must